detalles en show

+++ main.py
@@ -9,8 +9,9 @@ import find
 
 print "Welcome to py-note!\n"
 PATH = os.getcwd()
-arc = open("metadata", "w")
-arc.close()
+if not os.path.isfile("metadata"):
+    arc = open("metadata", "w")
+    arc.close()
 while True:
     input = raw_input("Command: ")
     if re.match("create",input) :
@@ -22,7 +23,7 @@ while True:
     elif re.match("edit", input):
         edit.op_edit(input)
     elif re.match("delete", input):
-        delete.op_delete(input)
+        delete.op_delete(input,PATH)
     elif re.match("find", input):
         find.op_find(input)
     elif input == "exit":
